+++ lib/features/clients/widgets/views/client_view_body.dart
@@ -76,6 +76,7 @@ class ClientDetailViewBody extends ConsumerWidget {
             child: WaveSyncBadge(
               syncState: client.waveSyncState,
               syncError: client.waveSyncError,
+              problems: client.waveProblems,
             ),
           ),
         ],